Step size of rating bar in review fragment is 1

+++ app/src/main/java/com/example/ratemovie/domain/entities/Review.kt
@@ -8,8 +8,8 @@ import kotlinx.parcelize.Parcelize
 @Parcelize
 data class Review(
     val text: String,
-    val grade: Float,
+    val grade: Int,
     val username: String,
 ) : Parcelable {
-    constructor() : this("", 0f, "")
+    constructor() : this("", 0, "")
 }
